requirements.txt
streamlit==1.37.1
pandas==2.2.2
pydantic==2.7.4
python-dotenv==1.0.1
openai==1.40.2
sqlalchemy==2.0.30
plotly==5.22.0
pdfplumber==0.10.3
reportlab==4.0.7

# Simple OCR Dependencies
pytesseract==0.3.10
Pillow==10.4.0
pdf2image==1.17.0
opencv-python==4.9.0.80
